+++ IngrediScan/Model/Local/ShoppingItem.swift
@@ -10,6 +10,6 @@ import Foundation
 struct ShoppingItem: Identifiable, Codable {
     var id = UUID()
     var ingredientId: Int
-    var amount: Int
+    var amount: Float
     var isBought: Bool = false
 }
